[FIX] Rename docx folder to doc
